Restructure Importer to allow different versions of MDES. Add Importer for MDES 2.0. #2574.
