+++ lib/ncs_navigator/staff_portal/warehouse/two_point_zero.rb
@@ -2,5 +2,6 @@ require 'ncs_navigator/staff_portal'
 module NcsNavigator::StaffPortal::Warehouse
   module TwoPointZero
     autoload :Enumerator, 'ncs_navigator/staff_portal/warehouse/two_point_zero/enumerator'
+    autoload :Importer, 'ncs_navigator/staff_portal/warehouse/two_point_zero/importer'
   end
 end
